Fixed a bug in display + new test reclaim from grand children
The code was puting a comma before the regions even if there was no td.

I wrote a test that creates a tree of 3 levels, each with three
children and revokes it.

The other test creates a line of children and passes around a carve
before it is revoked by td0 and then all children are revoked by td0.